Fix download pdf and send email

+++ app/Http/Controllers/OrderListItemsController.php
@@ -5,11 +5,15 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CreateOrderListItemRequest;
 use App\Http\Requests\UpdateOrderListItemRequest;
 use App\Item;
+use App\Mail\SupplierMail;
 use App\OrderList;
 use App\OrderListItem;
 use App\Supplier;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Mail;
+use PDF;
 
 class OrderListItemsController extends Controller
 {
@@ -92,8 +96,57 @@ class OrderListItemsController extends Controller
     public function delete(Request $request)
     {
         $orderListItem = OrderListItem::findOrFail($request->id);
-        if (Auth::user()->can('delete', $orderListItem))
+        if (Auth::check() && Auth::user()->can('delete', $orderListItem))
             $orderListItem->delete();
         return redirect()->back();
+    }
+
+    public function downloadPDF($orderListId)
+    {
+        $orderList = OrderList::findOrFail($orderListId);
+        if (Auth::check() && Auth::user()->can('download_pdf', $orderList))
+        {
+            $pdf = $this->generatePDF($orderListId, $orderList->note);
+            return $pdf->download($orderList->note . '.pdf');
+        }
+        return redirect()->back();
+    }
+
+    public function sendEmail($orderListId)
+    {
+        $orderList = OrderList::findOrFail($orderListId);
+        if (Auth::check() && Auth::user()->can('send_email', $orderList))
+        {
+            $supplierIds = DB::table('order_list_items')->select('supplier_id')->
+                where('order_list_id', $orderListId)->groupBy('supplier_id')->get();
+
+            foreach($supplierIds as $supplierId)
+            {
+                $supplier = Supplier::findOrFail($supplierId->supplier_id);
+                $pdf = $this->generatePDF($orderListId, $orderList->note, $supplier->id);
+                Mail::to($supplier->email)->send(new SupplierMail($pdf, $supplier->name));
+            }
+        }
+        return redirect()->back();
+    }
+
+    private function generatePDF($orderListId, $orderListName, $supplierId = null)
+    {
+        if (\is_null($supplierId))
+        {
+            $orderListItems = OrderListItem::where('order_list_id', $orderListId)->
+                with('item')->with('supplier')->orderBy('supplier_id')->orderBy('supplier_sort_order')->get();
+            return PDF::loadView('pdf.suppliers', ['order_list_name' => $orderListName,
+                'order_list_items' => $orderListItems]);
+        }
+        else
+        {
+            $supplier = Supplier::findOrFail($supplierId);
+            $orderListItems = OrderListItem::where('order_list_id', $orderListId)->
+                where('supplier_id', $supplierId)->with('item')->orderBy('supplier_sort_order')->get();
+            return PDF::loadView('pdf.supplier', ['order_list_name' => $orderListName,
+                'supplier_name' => $supplier->name,
+                'order_list_items' => $orderListItems]);
+        }
     }
 }
